+++ src/std/alloc/src/platform/mod.rs
@@ -1,0 +1,6 @@
+mod cp;
+mod windows;
+#[cfg(not(target_os = "windows"))]
+pub use crate::platform::cp::*;
+#[cfg(target_os = "windows")]
+pub use crate::platform::windows::*;